Completely restructured user creation by admin user. Both admin user and the database manager defer to a user factory for user creation. Reading user data from a file is no longer a function template. When we want to read data from a file, we pass the user type of the data being read and the file path.
