+++ Debug/C++BC_GamePublishingPlatform.log
@@ -1,2 +1,7 @@
-﻿  MenuSystem.cpp
+﻿  AdminUser.cpp
+  DatabaseManager.cpp
+  MainApplication.cpp
+  MenuSystem.cpp
+  Users.cpp
+  Generating Code...
   C++BC_GamePublishingPlatform.vcxproj -> C:\Users\Bear\Documents\Assignment 2\Vapor\CPPBC_Vapor\Debug\C++BC_GamePublishingPlatform.exe
