+++ MainApplication.cpp
@@ -1,6 +1,7 @@
 //C++ Boot Camp - Task 2 - 2018-19 
 //Name: Ismail Movahedi
 //Student number: 28039547
+#pragma once
 #include <iostream>
 
 #include "DatabaseManager.h"
